Move scrub to transport integration
This allows each transport to scrub it's own idiosyncracies
