+++ src/ServiceControl.AcceptanceTesting/InfrastructureConfig/ConfigureEndpointAzureStorageQueueTransport.cs
@@ -31,5 +31,6 @@
         public string Name => TransportNames.AzureStorageQueue;
         public string TypeName => $"{typeof(ASQTransportCustomization).AssemblyQualifiedName}";
         public string ConnectionString { get; set; }
+        public string ScrubPlatformConnection(string input) => input;
     }
 }
