+++ src/ServiceControl.AcceptanceTesting/InfrastructureConfig/ConfigureEndpointSqlServerTransport.cs
@@ -1,6 +1,7 @@
 ﻿namespace ServiceControl.AcceptanceTesting.InfrastructureConfig
 {
     using System;
+    using System.Data.Common;
     using System.Data.SqlClient;
     using System.Linq;
     using System.Threading.Tasks;
@@ -45,6 +46,27 @@
         public string Name => TransportNames.SQLServer;
         public string TypeName => $"{typeof(SqlServerTransportCustomization).AssemblyQualifiedName}";
         public string ConnectionString { get; set; }
+
+        public string ScrubPlatformConnection(string input)
+        {
+            var builder = new DbConnectionStringBuilder { ConnectionString = ConnectionString };
+
+            var result = input;
+
+            if (builder.TryGetValue("Database", out var database))
+            {
+                var databaseAsString = (string)database;
+                if (!string.IsNullOrEmpty(databaseAsString))
+                {
+                    result = result.Replace(
+                        $"[{databaseAsString}]",
+                        "[DATABASE]"
+                    );
+                }
+            }
+
+            return result;
+        }
 
         static async Task TryDeleteTable(SqlConnection conn, QueueAddress address)
         {
